Add create pairing while restricting members
The create pair with restricted button will split the team members
into two groups, restricted and pairable. It will then pair a random
restricted member with a random pairable member. If either of those
groups is empty, a warning message will be delivered. If there are no
available pairs to be created the button will be disabled.